Move operations to .types.operations
Move the classes under com.expediagroup.graphql.types.server.operations to com.expedigroup.graphql.types.operations
This moves everything under the graphql-kotlin-types module to the .types. package
