+++ examples/federation/base-app/src/main/kotlin/com/expediagroup/graphql/examples/query/NestedQuery.kt
@@ -16,7 +16,7 @@
 
 package com.expediagroup.graphql.examples.query
 
-import com.expediagroup.graphql.server.operations.Query
+import com.expediagroup.graphql.types.operations.Query
 import org.springframework.stereotype.Component
 import kotlin.random.Random
