+++ examples/server/spring-server/src/main/kotlin/com/expediagroup/graphql/examples/mutation/OptionalInputMutation.kt
@@ -16,7 +16,7 @@
 
 package com.expediagroup.graphql.examples.mutation
 
-import com.expediagroup.graphql.server.operations.Mutation
+import com.expediagroup.graphql.types.operations.Mutation
 import graphql.schema.DataFetchingEnvironment
 import org.springframework.stereotype.Component
