+++ examples/server/spring-server/src/main/kotlin/com/expediagroup/graphql/examples/query/EnvironmentQuery.kt
@@ -16,7 +16,7 @@
 
 package com.expediagroup.graphql.examples.query
 
-import com.expediagroup.graphql.server.operations.Query
+import com.expediagroup.graphql.types.operations.Query
 import graphql.schema.DataFetchingEnvironment
 import org.springframework.stereotype.Component
