+++ examples/server/spring-server/src/main/kotlin/com/expediagroup/graphql/examples/query/ScalarQuery.kt
@@ -17,8 +17,8 @@
 package com.expediagroup.graphql.examples.query
 
 import com.expediagroup.graphql.annotations.GraphQLDescription
-import com.expediagroup.graphql.server.operations.Mutation
-import com.expediagroup.graphql.server.operations.Query
+import com.expediagroup.graphql.types.operations.Mutation
+import com.expediagroup.graphql.types.operations.Query
 import com.expediagroup.graphql.scalars.ID
 import org.springframework.stereotype.Component
 import java.math.BigDecimal
