+++ examples/server/spring-server/src/main/kotlin/com/expediagroup/graphql/examples/subscriptions/SimpleSubscription.kt
@@ -19,7 +19,7 @@ package com.expediagroup.graphql.examples.subscriptions
 import com.expediagroup.graphql.annotations.GraphQLDescription
 import com.expediagroup.graphql.examples.context.MySubscriptionGraphQLContext
 import com.expediagroup.graphql.server.exception.KotlinGraphQLError
-import com.expediagroup.graphql.server.operations.Subscription
+import com.expediagroup.graphql.types.operations.Subscription
 import graphql.execution.DataFetcherResult
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.reactive.asPublisher
