+++ plugins/graphql-kotlin-gradle-plugin/src/test/resources/templates/HelloWorldQuery.mustache
@@ -1,6 +1,6 @@
 package com.example
 
-import com.expediagroup.graphql.server.operations.Query
+import com.expediagroup.graphql.types.operations.Query
 import org.springframework.stereotype.Component
 {{#customScalarsEnabled}}
 import java.util.UUID
