+++ plugins/graphql-kotlin-maven-plugin/src/integration/generate-sdl-federated/src/main/kotlin/com/example/SimpleQuery.kt
@@ -1,6 +1,6 @@
 package com.example
 
-import com.expediagroup.graphql.server.operations.Query
+import com.expediagroup.graphql.types.operations.Query
 import org.springframework.stereotype.Component
 
 @Component
