+++ plugins/schema/graphql-kotlin-sdl-generator/src/test/kotlin/com/expediagroup/test/SimpleQuery.kt
@@ -16,7 +16,7 @@
 
 package com.expediagroup.test
 
-import com.expediagroup.graphql.server.operations.Query
+import com.expediagroup.graphql.types.operations.Query
 import org.springframework.stereotype.Component
 
 @Component
