+++ servers/graphql-kotlin-spring-server/src/main/kotlin/com/expediagroup/graphql/server/spring/NonFederatedSchemaAutoConfiguration.kt
@@ -22,9 +22,9 @@ import com.expediagroup.graphql.execution.KotlinDataFetcherFactoryProvider
 import com.expediagroup.graphql.extensions.print
 import com.expediagroup.graphql.hooks.NoopSchemaGeneratorHooks
 import com.expediagroup.graphql.hooks.SchemaGeneratorHooks
-import com.expediagroup.graphql.server.operations.Mutation
-import com.expediagroup.graphql.server.operations.Query
-import com.expediagroup.graphql.server.operations.Subscription
+import com.expediagroup.graphql.types.operations.Mutation
+import com.expediagroup.graphql.types.operations.Query
+import com.expediagroup.graphql.types.operations.Subscription
 import com.expediagroup.graphql.server.spring.extensions.toTopLevelObjects
 import com.expediagroup.graphql.toSchema
 import graphql.schema.GraphQLSchema
