+++ servers/graphql-kotlin-spring-server/src/test/kotlin/com/expediagroup/graphql/server/spring/instrumentation/InstrumentationIT.kt
@@ -16,7 +16,7 @@
 
 package com.expediagroup.graphql.server.spring.instrumentation
 
-import com.expediagroup.graphql.server.operations.Query
+import com.expediagroup.graphql.types.operations.Query
 import com.expediagroup.graphql.server.spring.DEFAULT_INSTRUMENTATION_ORDER
 import com.expediagroup.graphql.types.GraphQLRequest
 import graphql.ExecutionResult
